# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FarmApp API
  version: 1.1.0
  description: API za gamificirano ucenje farmakologije na hrvatskom jeziku.
servers:
  - url: /api
paths:
...
  /players:
    post:
      operationId: createPlayer
      summary: Create or reuse a player profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "200":
          description: Existing player returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "201":
          description: New player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
  /topics:
    get:
      operationId: getTopics
      summary: Get topic summaries
      parameters:
        - in: query
          name: playerId
          required: false
          schema:
            type: string
        - in: query
          name: sourceImportId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Topic summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopicSummary"
  /dashboard/{playerId}:
    get:
      operationId: getDashboard
      summary: Get dashboard data for a player
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /quiz/next:
    get:
      operationId: getNextQuizCard
      summary: Get the next quiz card
      parameters:
        - in: query
          name: playerId
          required: true
          schema:
            type: string
        - in: query
          name: topicId
          required: false
          schema:
            type: string
        - in: query
          name: sourceImportId
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Next question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NextQuizCard"
  /quiz/answer:
    post:
      operationId: submitQuizAnswer
      summary: Submit an answer to a quiz card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAnswerRequest"
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizAnswerResult"
...
  /imports:
    get:
      operationId: getImports
      summary: Get available import collections
      responses:
        "200":
          description: Import summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceImportSummary"
    post:
      operationId: createImport
      summary: Publish a new question bank import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPayloadRequest"
      responses:
        "201":
          description: Import published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreview"
  /imports/{importId}:
    get:
      operationId: getImportDetail
      summary: Get import detail with lessons and topics
      parameters:
        - in: path
          name: importId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Import detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportDetail"
    patch:
      operationId: updateImport
      summary: Rename a subject or toggle its star state
      parameters:
        - in: path
          name: importId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateImportRequest"
      responses:
        "200":
          description: Updated import summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceImportSummary"
    delete:
      operationId: deleteImport
      summary: Archive a subject so it no longer appears in the library
      parameters:
        - in: path
          name: importId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Import archived
  /imports/preview:
    post:
      operationId: previewImport
      summary: Preview a question bank import before publishing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPayloadRequest"
      responses:
        "200":
          description: Import preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreview"
components:
  schemas:
...
    CreatePlayerRequest:
      type: object
      required:
        - nickname
      properties:
        nickname:
          type: string
    PlayerProfile:
      type: object
      required:
        - id
        - nickname
        - xp
        - level
        - streak
        - hearts
        - totalAnswered
        - correctAnswers
        - accuracy
      properties:
        id:
          type: string
        nickname:
          type: string
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        hearts:
          type: integer
        totalAnswered:
          type: integer
        correctAnswers:
          type: integer
        accuracy:
          type: integer
    TopicSummary:
      type: object
      required:
        - id
        - slug
        - titleHr
        - descriptionHr
        - accent
        - difficulty
        - questionCount
        - mastery
        - sourceLabelHr
      properties:
        id:
          type: string
        slug:
          type: string
        titleHr:
          type: string
        descriptionHr:
          type: string
        accent:
          type: string
        difficulty:
          type: string
          enum:
            - EASY
            - MEDIUM
            - HARD
        questionCount:
          type: integer
        mastery:
          type: integer
        sourceLabelHr:
          type: string
...
    WeakTopic:
      type: object
      required:
        - titleHr
        - accuracy
        - tipHr
      properties:
        titleHr:
          type: string
        accuracy:
          type: integer
        tipHr:
          type: string
    Dashboard:
      type: object
      required:
        - player
        - topics
        - leaderboard
        - imports
        - dailyMission
        - weakTopic
      properties:
        player:
          $ref: "#/components/schemas/PlayerProfile"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/TopicSummary"
        leaderboard:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        imports:
          type: array
          items:
            $ref: "#/components/schemas/SourceImportSummary"
        dailyMission:
          $ref: "#/components/schemas/DailyMission"
        weakTopic:
          $ref: "#/components/schemas/WeakTopic"
...
    NextQuizCard:
      type: object
      required:
        - flashcardId
        - topicId
        - topicTitleHr
        - questionHr
        - options
        - progressLabel
        - difficulty
      properties:
        flashcardId:
          type: string
        topicId:
          type: string
        topicTitleHr:
          type: string
        questionHr:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
        progressLabel:
          type: string
        difficulty:
          type: string
          enum:
            - EASY
            - MEDIUM
            - HARD
    QuizAnswerRequest:
      type: object
      required:
        - playerId
        - flashcardId
        - selectedOptionId
      properties:
        playerId:
          type: string
        flashcardId:
          type: string
        selectedOptionId:
          type: string
    QuizAnswerResult:
      type: object
      required:
        - isCorrect
        - correctOptionId
        - explanationHr
        - xpEarned
        - level
        - xp
        - streak
        - hearts
        - nextUnlockHr
        - topicMastery
      properties:
        isCorrect:
          type: boolean
        correctOptionId:
          type: string
        explanationHr:
          type: string
        xpEarned:
          type: integer
        level:
          type: integer
        xp:
          type: integer
        streak:
          type: integer
        hearts:
          type: integer
        nextUnlockHr:
          type: string
        topicMastery:
          type: integer
    LeaderboardEntry:
      type: object
      required:
        - nickname
        - level
        - xp
        - streak
      properties:
        nickname:
          type: string
        level:
          type: integer
        xp:
          type: integer
        streak:
          type: integer
    SourceImportSummary:
      type: object
      required:
        - id
        - titleHr
        - sourceType
        - status
        - isPinned
        - isSystem
        - createdAt
        - topicCount
        - lessonCount
        - questionCount
        - processingMode
        - processingNotesHr
      properties:
        id:
          type: string
        titleHr:
          type: string
        sourceType:
          type: string
          enum:
            - SCRIPT
            - JSON
            - PDF
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
        isPinned:
          type: boolean
        isSystem:
          type: boolean
        createdAt:
          type: string
          format: date-time
        topicCount:
          type: integer
        lessonCount:
          type: integer
        questionCount:
          type: integer
        processingMode:
          type: string
          enum:
            - AI
            - HEURISTIC
        processingNotesHr:
          type: string
    ImportPayloadRequest:
      type: object
      required:
        - titleHr
      properties:
        titleHr:
          type: string
        payloadText:
          type: string
        originalFileName:
          type: string
        pdfFileData:
          type: string
        documentPageCount:
          type: integer
        documentByteSize:
          type: integer
        sourceTypeHint:
          type: string
          enum:
            - SCRIPT
            - JSON
            - PDF
    UpdateImportRequest:
      type: object
      properties:
        titleHr:
          type: string
        isPinned:
          type: boolean
...
    ImportPreview:
      type: object
      required:
        - titleHr
        - sourceType
        - processingMode
        - processingNotesHr
        - topicCount
        - questionCount
        - topics
        - lessons
        - sampleQuestions
      properties:
        importId:
          type: string
        titleHr:
          type: string
        sourceType:
          type: string
          enum:
            - SCRIPT
            - JSON
            - PDF
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
        isPinned:
          type: boolean
        isSystem:
          type: boolean
        processingMode:
          type: string
          enum:
            - AI
            - HEURISTIC
        processingNotesHr:
          type: string
        topicCount:
          type: integer
        questionCount:
          type: integer
        topics:
          type: array
          items:
            $ref: "#/components/schemas/ImportPreviewTopic"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/StudyLesson"
        sampleQuestions:
          type: array
          items:
            $ref: "#/components/schemas/SampleQuestionPreview"
    ImportDetail:
      type: object
      required:
        - import
        - lessons
        - topics
      properties:
        import:
          $ref: "#/components/schemas/SourceImportSummary"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/StudyLesson"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/TopicSummary"